% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clim_data.R
\name{clim_data}
\alias{clim_data}
\title{clim_data()
Crops the spatial polygon of interest on global raster data;
compute the 19 annual bioclimatic variables for that polygon;
(intermediate files are exported to local memory here)
average these variables over defined periods of time.
(final files are exported to local memory here).
Libraries needed:
c(raster, dismo, rgdal, dplyr)}
\usage{
clim_data(
  data_dir = "./raw climatic data/",
  ex,
  cuts,
  year = 1961:2018,
  out_crop_dir = "./cropped climatic data/",
  out_average_dir = "./averaged and cropped climatic data/",
  print_stack_per_list = F
)
}
\arguments{
\item{ex}{= vector of geographical limits (W, E, S, N)}

\item{cuts}{= vector of years to produce the cuts for period averaging. c(1990) separates (earliest_year, 1989) and (1990, latest_year) periods}

\item{year}{= vector giving listing the years to consider (from 1961 to 2018 by default)}

\item{out_crop_dir}{= directory where to export geographically cropped bioclimatic data}

\item{out_average_dir}{= directory where to export averaged bioclimatic data}

\item{print_stack_per_list}{= (FALSE by default) if TRUE, prints final list of stacks}
}
\description{
clim_data()
Crops the spatial polygon of interest on global raster data;
compute the 19 annual bioclimatic variables for that polygon;
(intermediate files are exported to local memory here)
average these variables over defined periods of time.
(final files are exported to local memory here).
Libraries needed:
c(raster, dismo, rgdal, dplyr)
}
